Endpoints quickfix

+++ routes/tickets.js
@@ -57,6 +57,11 @@ router.post("/", async (req, res) => {
 });
 
 router.put("/:id", validatePut, async (req, res) => {
+  if (req.user.role !== "Aula") {
+    return res.status(401).json({
+      msg: "No tienes permisos para actualizar tickets",
+    });
+  }
    try {
       const {id} = req.params;
       const result = await updateTicket(id, req.body);
@@ -71,6 +76,11 @@ router.put("/:id", validatePut, async (req, res) => {
 });
 
 router.delete("/:id", async (req, res) => {
+  if (req.user.role !== "Aula") {
+    return res.status(401).json({
+      msg: "No tienes permisos para actualizar tickets",
+    });
+  }
    try {
       const {id} = req.params;
       const result = await deleteTicket(id);
